Got AD9833LIB working with SEEED2040 dev board
Had to change AD9833.h global array defs to extern
and also define these globals in main.  Didn't have
to do that last time I worked with the lib.  Can't
find a smoking gun but I think it's due to changes
to CMAKE, which was recently updated on my Ubuntu system.